Scroll a menu list with the wheel; the page is the byte hover never writes
A list longer than its window could only be paged with the D-pad, so three
of a seven-entry inventory's rows were unreachable by pointing (measured,
`7 entries, 4 visible`). The first version of the patch had a wheel and took
it out for a good reason -- it stepped the cursor, relative to where the
cursor was, while hover puts the cursor where the pointer is, and the two
fought on the next iteration of the menu loop.

The scrolling list has a second axis and it is the way out. `+0x20` is the
page and `+0x21` is the selection, and hover has never written the first of
them, because a pointer can only reach a row that is drawn. So the wheel
takes the page and hover keeps the cursor: scrolling under a still pointer
changes which entries the rows show, hover reads off the row the pointer is
on, and neither order disagrees. `+0x22` is rewritten for a page that moved
under a cursor that did not, since it is `+0x21` minus `+0x20`. Clamped to
`count - visible` rather than wrapped like the game's own arm, one notch to
one row, and the scrolling list's alone -- a fixed list and a prompt draw
every row they have, so a notch over one is discarded rather than saved.

0038 takes the notch off the host rather than a listener, in the drained
shape 0017's TakeMouseMotion already has: scroll arrives as discrete events,
so ImGui's per-frame level would lose one whenever two frames passed between
menu iterations, and a MouseEvent listener would have to be scoped to a
session that not every scrolling list is inside -- the save-slot menu comes
off the object-use handler and a shop off an NPC. Float end to end;
`(int)wheel.Y` rounded a trackpad's sub-notch scroll away entirely.

The gesture itself is unmeasured and cannot be measured from here: the shell
reaches Cross but not the menu's Up/Down and cannot inject a scroll at all,
so the arithmetic is read off func_8001EB70's own six branches, which are
now written down in docs/GAME_INTERNALS.md.

+++ tools/RecompOne/RecompOne.Runtime/Events/RuntimeEvents/MouseEvent.cs
@@ -24,5 +24,9 @@ public sealed class MouseEvent : IEvent
     public int X, Y;
     public MouseButton Button = MouseButton.None;
     public bool Pressed;
-    public int Wheel;
+    /// <summary>Notches scrolled, positive away from the user. A float for the
+    /// reason <c>InputManager.TakeMouseWheel</c> is one: a trackpad delivers a
+    /// fraction of a notch at a time, and an int throws every one of them
+    /// away.</summary>
+    public float Wheel;
 }
